LayoutTests/fast/css/image-set-parsing.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Test the parsing of the image-set and -webkit-image-set functions.");

function serializedRule(property, value)
{
    var div = document.createElement("div");
    div.setAttribute("style", property + ": " + value);
    document.body.appendChild(div);
    var result = div.style.getPropertyValue(property);
    document.body.removeChild(div);
    return result;
}

function testImageSetRule(description, property, rule, expectedSerialization)
{
    ["image-set", "-webkit-image-set"].forEach((valueName) => {
        debug("");
        debug(`${description}`);
        shouldBeEqualToString("serializedRule('" + property + "', '" + valueName + "(" + rule.replace(/'/g, "\\'") + ")')", `image-set(${expectedSerialization})`);

    });
}

testImageSetRule("Single value for -webkit-mask-box-image-source",
                "-webkit-mask-box-image-source",
                "url('#a')",
                "url(\"#a\") 1x");

testImageSetRule("Multiple values for -webkit-mask-box-image-source",
                "-webkit-mask-box-image-source",
                "url('#a'), url('#b') 2x",
                "url(\"#a\") 1x, url(\"#b\") 2x");

testImageSetRule("Single value for -webkit-mask-box-image-source with gradient.",
                "-webkit-mask-box-image-source",
                "linear-gradient(green, white)",
                "linear-gradient(green, white) 1x");

testImageSetRule("Multiple values for -webkit-mask-box-image-source with gradients",
                "-webkit-mask-box-image-source",
                "linear-gradient(green, white) 1x, radial-gradient(blue, white) 2x",
                "linear-gradient(green, white) 1x, radial-gradient(blue, white) 2x");

testImageSetRule("Combined gradient and URL for -webkit-mask-box-image-source",
                "-webkit-mask-box-image-source",
                "url(\"#a\"), linear-gradient(green, white) 2x",
                "url(\"#a\") 1x, linear-gradient(green, white) 2x");

testImageSetRule("Single -webkit-cross-fade for -webkit-mask-box-image-source",
                "-webkit-mask-box-image-source",
                "-webkit-cross-fade(url(\"#a\"), url(\"#b\"), 70%)",
                "-webkit-cross-fade(url(\"#a\"), url(\"#b\"), 0.7) 1x");

testImageSetRule("Combined gradient and cross-fade for -webkit-mask-box-image-source",
                "-webkit-mask-box-image-source",
                "-webkit-cross-fade(url(\"#a\"), url(\"#b\"), 70%), linear-gradient(green, white) 2x",
                "-webkit-cross-fade(url(\"#a\"), url(\"#b\"), 0.7) 1x, linear-gradient(green, white) 2x");

testImageSetRule("Single value for -webkit-mask-box-image-source with gradient and dpi",
                "-webkit-mask-box-image-source",
                "linear-gradient(green, white) 192dpi",
                "linear-gradient(green, white) 192dpi");

successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
